+++ externs/twos-complement-buffer.js
@@ -52,8 +52,8 @@ TwosComplementBuffer.prototype.bytes = 0;
  * @param {number} num The number.
  * @param {number=} index The index being written in the byte buffer.
  * @return {number} The next index to write on the byte buffer.
- * @throws {Error} If num is NaN.
- * @throws {Error} On overflow.
+ * @throws {TypeError} If num is NaN.
+ * @throws {RangeError} On overflow.
  */
 TwosComplementBuffer.prototype.pack = function(buffer, num, index=0) {};
 
@@ -62,6 +62,6 @@ TwosComplementBuffer.prototype.pack = function(buffer, num, index=0) {};
  * @param {!Uint8Array|!Array<number>} buffer An array of bytes.
  * @param {number=} index The index to read.
  * @return {number}
- * @throws {Error} On overflow.
+ * @throws {RangeError} On overflow.
  */
 TwosComplementBuffer.prototype.unpack = function(buffer, index=0) {};