+++ test/src/int16.js
@@ -48,25 +48,25 @@ describe('16-bit integers, signed', function() {
     assert.throws(function(){
       var buffer = [];
       tcb.pack(buffer, -32769); // ommit the index, ==0
-    }, /Overflow/);
+    }, RangeError);
   });
   it('throw error if pack 32768', function() {
     assert.throws(function(){
       var buffer = [];
       tcb.pack(buffer, 32768); // ommit the index, ==0
-    }, /Overflow/);
+    }, RangeError);
   });
   it('throw error if pack -Infinity', function() {
     assert.throws(function(){
       var buffer = [];
       tcb.pack(buffer, -Infinity); // ommit the index, ==0
-    }, /Overflow/);
+    }, RangeError);
   });
   it('throw error if pack Infinity', function() {
     assert.throws(function(){
       var buffer = [];
       tcb.pack(buffer, Infinity); // ommit the index, ==0
-    }, /Overflow/);
+    }, RangeError);
   });
 
   // NaN
@@ -74,6 +74,6 @@ describe('16-bit integers, signed', function() {
     assert.throws(function(){
       var buffer = [];
       tcb.pack(buffer, NaN); // ommit the index, ==0
-    }, /NaN/);
+    }, TypeError);
   });
 });